# docs/games-api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WinX Games API",
    "version": "1.0.0",
    "description": "Games module endpoints. No global prefix; all routes under /games. Responses are wrapped as { statusCode, message, data }."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "required": ["intervalMinutes", "betDurationMinutes", "drawOffsetMinutes", "roundCodePrefix"],
        "properties": {
          "intervalMinutes": { "type": "integer", "minimum": 1, "maximum": 1440 },
          "betDurationMinutes": { "type": "integer", "minimum": 1, "maximum": 1440, "description": "Must be < intervalMinutes" },
          "drawOffsetMinutes": { "type": "integer", "minimum": 0, "maximum": 60 },
          "roundCodePrefix": { "type": "string", "maxLength": 20 },
          "nextRunAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" }
        }
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "intervalMinutes": { "type": "integer", "minimum": 1, "maximum": 1440 },
          "betDurationMinutes": { "type": "integer", "minimum": 1, "maximum": 1440 },
          "drawOffsetMinutes": { "type": "integer", "minimum": 0, "maximum": 60 },
          "roundCodePrefix": { "type": "string", "maxLength": 20 },
          "nextRunAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" }
        }
      },
...
      "PlaceBet": {
        "type": "object",
        "required": ["game_id", "round_id", "bet_number", "amount"],
        "properties": {
          "user_id": { "type": "integer", "description": "Required for /games/bet; ignored for /games/bet-with-ticket (taken from token)" },
          "game_id": { "type": "integer" },
          "round_id": { "type": "integer" },
          "bet_number": { "type": "string" },
          "amount": { "type": "number" }
        }
      },
...
    }
  },
  "paths": {
...
    "/games": {
      "get": {
        "tags": ["Public"], "summary": "List games (filterable)",
        "parameters": [
          { "name": "isActive", "in": "query", "schema": { "type": "boolean" } },
          { "name": "isHot", "in": "query", "schema": { "type": "boolean" } },
          { "name": "isJackpotBadge", "in": "query", "schema": { "type": "boolean" } },
          { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["REGULAR", "JACKPOT", "INSTANT", "CUSTOM"] } },
          { "name": "digitLength", "in": "query", "schema": { "type": "integer", "enum": [1, 3, 4, 5] } },
          { "name": "liveOnly", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/games/bet-with-ticket": {
      "post": {
        "tags": ["User"], "summary": "Place bet + generate PDF ticket", "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlaceBet" } } } },
        "responses": { "201": { "description": "Created" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/games/admin/schedules": {
      "get": { "tags": ["Admin - Schedules"], "summary": "List all schedules", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/games/admin/{gameId}/schedule": {
      "get": {
        "tags": ["Admin - Schedules"], "summary": "Get a game's schedule", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "gameId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "No schedule" } }
      },
      "post": {
        "tags": ["Admin - Schedules"], "summary": "Create schedule", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "gameId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateScheduleDto" } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "betDuration >= interval or duplicate" } }
      },
      "patch": {
        "tags": ["Admin - Schedules"], "summary": "Update schedule (partial)", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "gameId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateScheduleDto" } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Validation error" } }
      },
      "delete": {
        "tags": ["Admin - Schedules"], "summary": "Delete schedule", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "gameId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/games/admin/{gameId}/schedule/toggle": {
      "post": {
        "tags": ["Admin - Schedules"], "summary": "Pause/resume schedule", "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "gameId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}